Fix destroy with non-matched id's

+++ lib/AdoDBRecord/Base.class.php
@@ -88,7 +88,7 @@
 				return;
 			}
 			$obj = AdoDBRecord_Base::find($id);
-			$obj->delete();
+			if (is_object($obj)) $obj->delete();
 		}
 
 		# returns the records found by $arguments
